capturing errors in an array instead of stubbing behaviour
Needs more investigations, but here we are using a completions array to capture errors and test the contents of the array in the unit test.

This is done so we don't have to stub error for the mock class but imo it still seems like we are kinda stubbing with a closure. complete(with:..) function remedies this a bit but still need practice with this approach