Refactoring to take statusCode into account
In previous tests we were defining empty data and disregarding status code in the load method. Now we inject the json Data using the makeItemsJson method and check status code of response

Earlier we could send a wrong response code and valid json and have the tests pass but this is not what we want. We want both a valid json and 200 status code.